Finished up Fragment Navigation + UI
I locked in and pieced everything together as it needed to be. I relied on ChatGPT to help me refactor Demo Code for the specifications of this assignment. I also used it to help make my layout files look nice with the necessary inputs and elements. I then went through each file and made sure the code was doing what it needed to and nothing more. The data is properly setup with the RecyclerView and making use of the MutableStateFlow. The MainActivity is the only Activity and simply hosts the ContactListFragment on launch. The Fragments properly navigate to and from each other and support item deletion. ChatGTP helped me figure out how to pass the arguments to the Details Fragment so I could display the correct contact.